Refactoring Hymod package organization and file structure. (#100)
* Move HymodErrorCodes enum to dedicated file.

* Move hymod_params struct to dedicated file.

* Move hymod_state struct to dedicated file.

* Move hymod_fluxes struct to dedicated file.

* Remove pass-thru stub func in Hymod for ET.

Removing previous calc_et function in hymod_kernel class, which only
returned 0.0, renaming the previous calc_et2 to calc_et, and commenting
out unit test for calc_et until we have real samples to test against.
