+++ models/hymod/include/Hymod.h
@@ -3,100 +3,24 @@
 
 #include <cmath>
 #include <vector>
+#include "HymodErrorCodes.h"
 #include "LinearReservoir.hpp"
 #include "Nonlinear_Reservoir.hpp"
 #include "Pdm03.h"
-
-//! Hymod paramaters struct
-/*!
-    This structure provides storage for the parameters of the hymod hydrological model
-*/
-
-struct hymod_params
-{
-    double max_storage_meters; //!< maximum amount of water stored
-    double a;               //!< coefficent for distributing runoff and slowflow
-    double b;               //!< exponent for flux equation
-    //Ks and Kq are coeeficint constants used by the non-linear reservoirs.  There is an implicit unit of time
-    //in these parameters to make `a*(dS/S)^b` have approriate units of meters/second
-    //this implies that for any given timstep, dt, the approriate coefficients may be different.
-    double Ks;              //!< slow flow coefficent
-    double Kq;              //!< quick flow coefficent
-    int n;               //!< number of nash cascades
-};
-
-//! Hymod state structure
-/*!
-    This structure provides storage for the state used by hymod hydrological model at a particular time step.
-    Warning: to allow bindings to other languages the storage amounts for the reserviors in the nash cascade
-    are stored in a pointer which is not allocated or deallocated by this code. Backing storage for these arrays
-    must be provide and managed by the creator of these structures.
-*/
-
-struct hymod_state
-{
-    double storage_meters;             //!< the current water storage of the modeled area
-    double groundwater_storage_meters; //!< the current water in the ground water nonlinear reservoir
-    double* Sr;                 //!< amount of water in each nonlinear reservoir unsafe for binding suport check latter
-
-    //! Constructuor for hymod state
-    /*!
-        Default constructor for hymod_state objects. This is necessary for the structure to be usable in a map
-        Warning: the value of the Sr pointer must be set before this object is used by the hymod_kernel::run() or hymod()
-    */
-    hymod_state(double inital_storage_meters = 0.0, double gw_storage_meters = 0.0, double* storage_reservoir_ptr = 0x0) :
-        storage_meters(inital_storage_meters), groundwater_storage_meters(gw_storage_meters), Sr(storage_reservoir_ptr)
-    {
-
-    }
-};
-
-//! Hymod flux structure
-/*!
-    This structure provides storage for the fluxes generated by hymod at any time step
-*/
-
-struct hymod_fluxes
-{
-    double slow_flow_meters_per_second;  //!< The flow exiting slow flow at this time step
-    double runoff_meters_per_second;     //!< The caluclated runoff amount for this time step
-    double et_loss_meters;    //!< The amount of water lost to evapotranspiration
-
-    //! Constructor for hymod fluxes
-    /*!
-        Default constructor for hymod fluxes this is needed for the structs to be storeable in C++ containers.
-    */
-
-    hymod_fluxes(double sf = 0.0, double r = 0.0, double et = 0.0) :
-        slow_flow_meters_per_second(sf), runoff_meters_per_second(r), et_loss_meters(et)
-    {}
-};
-
-enum HyModErrorCodes
-{
-    NO_ERROR = 0,
-    MASS_BALANCE_ERROR = 100
-};
+#include "hymod_params.h"
+#include "hymod_state.h"
+#include "hymod_fluxes.h"
 
 //! Hymod kernel class
 /*!
     This class implements the hymod hydrological model
-
-
 */
-
 class hymod_kernel
 {
     public:
 
     //! stub function to simulate losses due to evapotransportation
     static double calc_et(double soil_m, void* et_params)
-    {
-        return 0.0;
-    }
-
-    //! stub function to simulate losses due to evapotransportation
-    static double calc_et2(double soil_m, void* et_params)
     {
         pdm03_struct* pdm = (pdm03_struct*) et_params;
         pdm->XHuz = soil_m;
