+++ models/hymod/include/hymod_fluxes.h
@@ -1,0 +1,25 @@
+#ifndef NGEN_HYMOD_FLUXES_H
+#define NGEN_HYMOD_FLUXES_H
+
+//! Hymod flux structure
+/*!
+    This structure provides storage for the fluxes generated by hymod at any time step
+*/
+
+struct hymod_fluxes
+{
+    double slow_flow_meters_per_second;  //!< The flow exiting slow flow at this time step
+    double runoff_meters_per_second;     //!< The caluclated runoff amount for this time step
+    double et_loss_meters;    //!< The amount of water lost to evapotranspiration
+
+    //! Constructor for hymod fluxes
+    /*!
+        Default constructor for hymod fluxes this is needed for the structs to be storeable in C++ containers.
+    */
+
+    hymod_fluxes(double sf = 0.0, double r = 0.0, double et = 0.0) :
+            slow_flow_meters_per_second(sf), runoff_meters_per_second(r), et_loss_meters(et)
+    {}
+};
+
+#endif //NGEN_HYMOD_FLUXES_H
